[wordpress] updated core to 5.0 & changed wp-db folder name

+++ wordpress/wp-config.php
@@ -118,16 +118,13 @@ if ( !defined('ABSPATH') )
 	define('ABSPATH', dirname(__FILE__) . '/');
 
 /** Delete database dir if is not local enviroment **/
-if ( !$localhost && is_dir('wp-database') )
+if ( !$localhost && is_dir('wp-db') )
 {
-	foreach(glob('wp-database/{,.}*', GLOB_BRACE) as $filename)
+	foreach(glob('wp-db/{,.}*', GLOB_BRACE) as $filename)
 	{
-		if(is_file($filename))
-		{
-			unlink($filename);
-		}
+		if(is_file($filename)) unlink($filename);
 	}
-	rmdir('wp-database');
+	rmdir('wp-db');
 }
 
 /** Sets up WordPress vars and included files. */
